refactor pipes code
