+++ pydoop/mapreduce/connections.py
@@ -16,65 +16,59 @@
 #
 # END_COPYRIGHT
 
-import sys
 import os
 import socket
-import logging
 
-from .text_streams import TextDownStreamAdapter, TextUpStreamAdapter
-from .binary_streams import BinaryDownStreamAdapter, BinaryUpStreamAdapter
-
-
-logging.basicConfig(level=logging.CRITICAL)
-LOGGER = logging.getLogger('connections')
+import pydoop.sercore as sercore
+from .binary_streams import BinaryProtocol
 
 
-BUF_SIZE = 128 * 1024
+class Connection(object):
 
-
-class Connections(object):
-
-    def __init__(self, cmd_stream, up_link):
-        self.cmd_stream = cmd_stream
-        self.up_link = up_link
+    def __init__(self):
+        self.downlink = None
 
     def close(self):
-        self.cmd_stream.close()
-        self.up_link.flush()
-        self.up_link.close()
+        if self.downlink:
+            self.downlink.close()
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, *args):
+        self.close()
 
 
-def open_playback_connections(cmd_file, out_file, auto_serialize=True):
-    in_stream = open(cmd_file, 'rb')
-    out_stream = open(out_file, 'wb')
-    cmd_stream = BinaryDownStreamAdapter(in_stream)
-    up_link = BinaryUpStreamAdapter(out_stream, auto_serialize=auto_serialize)
-    return Connections(cmd_stream, up_link)
+class NetworkConnection(Connection):
 
-
-def open_file_connections(istream=sys.stdin, ostream=sys.stdout):
-    return Connections(TextDownStreamAdapter(istream),
-                       TextUpStreamAdapter(ostream))
-
-
-class NetworkConnections(Connections):
-
-    def __init__(self, cmd_stream, up_link, sock, port):
-        self.logger = LOGGER.getChild('NetworkConnections')
-        super(NetworkConnections, self).__init__(cmd_stream, up_link)
-        self.socket = sock
+    def __init__(self, context, host, port, **kwargs):
+        super(NetworkConnection, self).__init__()
+        self.socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        self.socket.connect((host, port))
+        istream = sercore.FileInStream(self.socket)
+        ostream = sercore.FileOutStream(self.socket)
+        self.downlink = BinaryProtocol(istream, context, ostream, **kwargs)
 
     def close(self):
-        super(NetworkConnections, self).close()
-        self.socket.shutdown(socket.SHUT_RDWR)
+        super(NetworkConnection, self).close()
         self.socket.close()
 
 
-def open_network_connections(port, auto_serialize=True):
-    s = socket.socket()
-    s.connect(('localhost', port))
-    in_stream = os.fdopen(os.dup(s.fileno()), 'r', BUF_SIZE)
-    out_stream = os.fdopen(os.dup(s.fileno()), 'w', BUF_SIZE)
-    cmd_stream = BinaryDownStreamAdapter(in_stream)
-    up_link = BinaryUpStreamAdapter(out_stream, auto_serialize=auto_serialize)
-    return NetworkConnections(cmd_stream, up_link, s, port)
+class FileConnection(Connection):
+
+    def __init__(self, context, in_fn, out_fn, **kwargs):
+        super(FileConnection, self).__init__()
+        istream = sercore.FileInStream(in_fn)
+        ostream = sercore.FileOutStream(out_fn)
+        self.downlink = BinaryProtocol(istream, context, ostream, **kwargs)
+
+
+def get_connection(context, **kwargs):
+    port = os.getenv("mapreduce.pipes.command.port")
+    if port:
+        return NetworkConnection(context, "localhost", int(port), **kwargs)
+    in_fn = os.getenv("mapreduce.pipes.commandfile")
+    if in_fn:
+        out_fn = "%s.out" % in_fn
+        return FileConnection(context, in_fn, out_fn, **kwargs)
+    raise NotImplementedError  # TBD: text protocol (all kwargs ignored?)
